chore: refactor and fix tests
