+++ libraries/config/include/config/dag_config.hpp
@@ -7,7 +7,7 @@
 namespace taraxa {
 
 struct DagConfig {
-  u256 gas_limit = 0x989680;
+  uint64_t gas_limit = 0x989680;
 };
 Json::Value enc_json(DagConfig const& obj);
 void dec_json(Json::Value const& json, DagConfig& obj);
