+++ libraries/config/include/config/pbft_config.hpp
@@ -13,7 +13,7 @@ struct PbftConfig {
   uint32_t dag_blocks_size = 0;
   uint32_t ghost_path_move_back = 0;
   bool run_count_votes = false;
-  u256 gas_limit = 0x989680;
+  uint64_t gas_limit = 0x3938700;
 };
 Json::Value enc_json(PbftConfig const& obj);
 void dec_json(Json::Value const& json, PbftConfig& obj);
